Refactored just about everything.
