+++ README.md
@@ -1,0 +1,13 @@
+# BattleSnake-2020
+A BattleSnake creation by github.com/Kristian-D and github.com/Ross-Alexandra
+
+# Running the server
+In order to run the server you need to add the root of this project to your GOPATH
+environmental variables. ex, If you have cloned the repo at `~/repos/battlesnake-2020`
+then you would add `~/repos/battlesnake-2020` to your GOPATH.
+
+The server can then needs to be build. This can be done with
+```
+go build ./src/battlesnake.theserverproject.com/start
+```
+
