+++ src/battlesnake.theserverproject.com/insert_clever_name/go.mod
@@ -1,0 +1,3 @@
+module battlesnake.theserverproject.com/insert_clever_name/snakeserver
+
+go 1.14
